tests/test/backend/requirement.txt
fastapi
uvicorn
asyncpg
pydantic
python-dotenv
sqlalchemy
bcrypt
psycopg2-binary
